% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{sbifapi}
\alias{sbifapi}
\title{Generic function for the GET method of the API}
\usage{
sbifapi(what = "dolar", frame = "year",
  apikey = "724b35d182347cbcc3b2f45b7f3f9888232ec91b", formato = "JSON",
  callback = NULL, Year = lubridate::year(Sys.Date()), Month = NULL,
  Day = NULL, retry = 1)
}
\arguments{
\item{what}{Object to retrieve (see \link{Details})}

\item{frame}{For the particular date, before that date, or after that date}

\item{apikey}{API key}

\item{formato}{Format of the returning data, can be either JSON or XML}

\item{callback}{Callback function for JSON data}

\item{Year}{Year}

\item{Month}{Month}

\item{Day}{Day}

\item{retry}{Number of times to retry the call to the API}
}
\description{
Generic function for the GET method of the API
}
\details{
Currently, the available objects are
\itemize{
 \item{what}{dolar},
 \item frame fixed, before, after
 year, year/month, year/month/dias/, posteriores/year,
 posteriores/year/month, posteriores/year/month/day, and the same with
 anteriores
 \item
}
}

